libthread: fix interworking issue with ARM assembly
This should enable the tooling to interwork between the A32 and T32
code by marking the assembly function as a function with the .type
directive.

Note that .type could be used in other Linux targets also, but I am not
sure if there would be any benefit to that (it is conceivable that
marking functions might be helpful to some debugging tools that work
with executables, like GDB or Ghidra).

Some related TODOs: The ARM code could use STM and LDM to remove the
need for a lot of STRs and LDRs. Also ".syntax unified" should be used
for the ARM assembly, so it would be easier to choose between the
various ARM instruction sets. Also save the FP registers.

Fixes #59

Change-Id: I7c818a4947d8acfd66cc45ae942af172967e6cfd